Add development image Dockerfile and entrypoint script
Note that the dev dockerfile is a child of todobackend-base.
It will override the base entrypoint.

Note ordering of Docker commands in Dockerfile.
Add of source code is done last because that is the most active
image layer.  Making it last means above layers dont need to be
re-imaged when new image is being created.   Cached layers make
the build run fast.