Create docker infrastructure for executing build stage of CD pipeline
This change builds application artifacts
This builds Python wheels
This uses test environment
This creates snapshot of app code and app dependencies which are saved
in build cache.  This assures the build cache is a true representation
of application as it was tested.

More changes:
1. Add setup.py file and move requirements files here as install and
extras requires.
2. Change requirements*.txt contents
3. Change test.sh to use requirements from build cache first
4. Change Dockerfile to include a build cache volume
5. Add a builder service that uses the parent base entrypoint and
launches the wheel creation while using only existing packages in build
cache.