Update docker-compose to ensure app service runs when nginx is run.
Add app under links in nginx.
Test hitting the <appurl>:8000 to test application.
See that static css is not loaded.
Fix this by running:
$ docker-compose run --rm  app manage.py collectstatic --noinput
The command "manage.py collectstatic" is needed by django to populate
the static files.   The command is idempotent and can be run as much as
necessary.